refactor(ai-workflow-builder): Move triage agent to agents/ and fix ask_assistant scope
Move TriageAgent, tool definitions, and tests from assistant/ to agents/
to co-locate with other builder agents. Update ask_assistant description
to exclude debugging/error-fixing (builder responsibilities) and fix
posthog mock in builder.utils.test.ts.
