+++ packages/@n8n/ai-workflow-builder.ee/src/agents/build-workflow.tool.ts
