+++ packages/@n8n/ai-workflow-builder.ee/src/agents/test/triage.agent.test.ts
@@ -1,12 +1,13 @@
 import type { BaseChatModel } from '@langchain/core/language_models/chat_models';
 import { AIMessage } from '@langchain/core/messages';
 
-import type { ConversationEntry } from '../../code-builder/utils/code-builder-session';
-import type { StreamOutput } from '../../types/streaming';
-import type { ChatPayload } from '../../workflow-builder-agent';
-import type { AssistantHandler } from '../assistant-handler';
-import { TriageAgent } from '../triage-agent';
-import type { TriageAgentOutcome } from '../triage-agent';
+import type { AssistantHandler } from '@/assistant/assistant-handler';
+import type { ConversationEntry } from '@/code-builder/utils/code-builder-session';
+import type { StreamOutput } from '@/types/streaming';
+import type { ChatPayload } from '@/workflow-builder-agent';
+
+import { TriageAgent } from '../triage.agent';
+import type { TriageAgentOutcome } from '../triage.agent';
 
 function createMockPayload(
 	message = 'test message',
