+++ packages/@n8n/ai-workflow-builder.ee/src/assistant/index.ts
@@ -1,6 +1,4 @@
 export { AssistantHandler } from './assistant-handler';
-export { TriageAgent } from './triage-agent';
-export type { TriageAgentConfig, TriageAgentOutcome } from './triage-agent';
 export type {
 	AssistantContext,
 	AssistantResult,
