+++ packages/frontend/editor-ui/src/app/constants/experiments.ts
@@ -91,11 +91,7 @@ export const CODE_WORKFLOW_BUILDER_EXPERIMENT = createExperiment('071_coding_wor
 
 export const MERGE_ASK_BUILD_FF = '075_merge_ask_build';
 
-export const QUICK_CONNECT_EXPERIMENT = {
-	name: '073_quick_connect',
-	control: 'control',
-	variant: 'variant',
-};
+export const QUICK_CONNECT_EXPERIMENT = createExperiment('073_quick_connect');
 
 export const EXPERIMENTS_TO_TRACK = [
 	EXTRA_TEMPLATE_LINKS_EXPERIMENT.name,
